Update group metadata version on membership change [schema change]
This should cause permission changes to sync down to 5.0.

This also changes the group's version column to a larger integer. Since
there's no ?since for group metadata and the version is just a shorter
ETag, it previously wrapped around at 255. The syncing documentation
does technically say to check if the version number is different, not
higher, but since I didn't even code it that way in the client, we'll
just go with a normal monotonically incrementing version number.

ALTER TABLE `groups` CHANGE `version` `version` MEDIUMINT UNSIGNED NOT
NULL DEFAULT '1';

+++ misc/master.sql
@@ -93,7 +93,7 @@ CREATE TABLE `groups` (
   `hasImage` tinyint(1) unsigned NOT NULL DEFAULT '0',
   `dateAdded` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `dateModified` timestamp NOT NULL DEFAULT '0000-00-00 00:00:00',
-  `version` tinyint(3) unsigned NOT NULL DEFAULT '1',
+  `version` mediumint(8) unsigned NOT NULL DEFAULT '1',
   PRIMARY KEY (`groupID`),
   UNIQUE KEY `libraryID` (`libraryID`),
   UNIQUE KEY `slug` (`slug`)
